requirements.txt
google-cloud-pubsub==2.13.11
avro==1.11.1